+++ core/src/state/splitted_state.rs
@@ -7,28 +7,79 @@ use crate::{
   widget::{Render, RenderBuilder, Widget},
 };
 use ribir_algo::Sc;
-use rxrust::{
-  ops::box_it::BoxOp,
-  prelude::{BoxIt, ObservableExt},
-};
+use rxrust::{observable::ObservableExt, ops::box_it::BoxOp, prelude::BoxIt};
 use std::{
   cell::{Cell, RefMut},
   time::Instant,
 };
 
 /// A writer splitted writer from another writer, and has its own notifier.
-pub struct SplittedWriter<O, R, W> {
+pub struct SplittedWriter<O, R, W, C> {
   origin: O,
   map: R,
   mut_map: W,
+  checker: Sc<C>,
   notifier: Notifier,
   batched_modify: Sc<Cell<ModifyScope>>,
-  create_at: Instant,
   last_modified: Sc<Cell<Instant>>,
   ref_count: Sc<Cell<usize>>,
 }
 
-impl<O, R, W> Drop for SplittedWriter<O, R, W> {
+/// The SplitChecker will be bind to the SplittedWriter or SplittedReader.
+/// The method of is_valid will be call before the visit to the Value of
+/// StateReader or StateWriter, by method of read() and write().
+/// the default SplitChecker to the SplittedWriter is VolatileSplit, see
+/// [VolatileSplit], and you can also set to AlwaysValidSplit(see
+/// [AlwaysValidSplit]), or set to your custom checker by set_checker of
+/// StateWriter
+pub trait SplitChecker<P: ?Sized> {
+  fn is_valid<S: StateReader<Value = P>>(&self, src: &S) -> bool;
+}
+
+/// The Default SplitChecker. It will return false if the origin writer has been
+/// changed or if the origin writer is invalid, and the continue used will cause
+/// panic.
+/// We set the SplittedWriter default with VolatileSplit checker to prevent the
+/// unexpected bug if the SplittedWriter is mapped in not stable way, e.x: map
+/// by index to  vector
+/// ### Example of SplittedWriter to Vec item.
+/// the follow example will panic to exposed the unexpected usage to visit to
+/// user Tom.
+/// ``` rust
+/// use ribir_core::prelude::*;
+/// #[derive(Debug)]
+/// struct User {
+///   name: &'static str,
+///   //..
+/// }
+/// let src = Stateful::new(vec![User { name: "Lily" }, User { name: "Lucy" }, User { name: "Tom" }]);
+/// let lucy_idx = 1;
+/// let lucy = src.split_writer(move |v| &v[lucy_idx], move |v| &mut v[lucy_idx]);
+/// src.write().remove(0);
+/// println!("lucy's info {:?}", *lucy.read());
+/// ```
+pub struct VolatileSplit {
+  create_at: Instant,
+}
+
+impl Default for VolatileSplit {
+  fn default() -> Self { VolatileSplit { create_at: Instant::now() } }
+}
+
+impl<P: ?Sized> SplitChecker<P> for VolatileSplit {
+  fn is_valid<S: StateReader<Value = P>>(&self, p: &S) -> bool {
+    self.create_at > p.time_stamp() && p.is_valid()
+  }
+}
+
+/// The AlwaysValidSplit SplitChecker. It will always return true if the
+/// OriginWriter is valid.
+pub struct AlwaysValidSplit {}
+impl<P: ?Sized> SplitChecker<P> for AlwaysValidSplit {
+  fn is_valid<S: StateReader<Value = P>>(&self, p: &S) -> bool { p.is_valid() }
+}
+
+impl<O, R, W, C> Drop for SplittedWriter<O, R, W, C> {
   fn drop(&mut self) {
     if self.ref_count.get() == 1 {
       let mut notifier = self.notifier.clone();
@@ -42,11 +93,11 @@ impl<O, R, W> Drop for SplittedWriter<O, R, W> {
   }
 }
 
-pub struct SplittedReader<S, F> {
+pub struct SplittedReader<S, F, C> {
   origin: S,
   map: F,
+  checker: Sc<C>,
   notifier: Notifier,
-  create_at: Instant,
   last_modified: Sc<Cell<Instant>>,
 }
 
@@ -54,14 +105,11 @@ macro_rules! splitted_reader_impl {
   () => {
     type Value = V;
     type OriginReader = O;
-    type Reader = SplittedReader<O::Reader, R>;
+    type Reader = SplittedReader<O::Reader, R, C>;
 
     #[track_caller]
     fn read(&self) -> ReadRef<Self::Value> {
-      assert!(
-        self.create_at > self.origin.time_stamp(),
-        "A splitted reader is invalid because its origin state is modified after it created."
-      );
+      assert!(self.is_valid(), "A splitted reader is invalid.");
       ReadRef::map(self.origin.read(), &self.map)
     }
 
@@ -71,10 +119,13 @@ macro_rules! splitted_reader_impl {
         origin: self.origin.clone_reader(),
         map: self.map.clone(),
         notifier: self.notifier.clone(),
-        create_at: self.create_at,
+        checker: self.checker.clone(),
         last_modified: self.last_modified.clone(),
       }
     }
+
+    #[inline]
+    fn is_valid(&self) -> bool { self.checker.is_valid(self.origin_reader()) }
 
     #[inline]
     fn origin_reader(&self) -> &Self::OriginReader { &self.origin }
@@ -84,12 +135,11 @@ macro_rules! splitted_reader_impl {
 
     #[inline]
     fn raw_modifies(&self) -> BoxOp<'static, ModifyScope, std::convert::Infallible> {
-      let origin = self.origin.clone_reader();
-      let create_at = self.create_at;
+      let this = self.clone_reader();
       self
         .notifier
         .raw_modifies()
-        .take_while(move |_| origin.time_stamp() < create_at)
+        .filter(move |_| this.is_valid())
         .box_it()
     }
 
@@ -103,17 +153,18 @@ macro_rules! splitted_reader_impl {
   };
 }
 
-impl<V, O, R> StateReader for SplittedReader<O, R>
+impl<V, O, R, C> StateReader for SplittedReader<O, R, C>
 where
   Self: 'static,
   V: ?Sized,
   O: StateReader,
   R: Fn(&O::Value) -> &V + Clone,
+  C: SplitChecker<O::Value>,
 {
   splitted_reader_impl!();
 }
 
-impl<V, O, R, W> StateReader for SplittedWriter<O, R, W>
+impl<V, O, R, W, C> SplittedWriter<O, R, W, C>
 where
   Self: 'static,
   V: ?Sized,
@@ -121,18 +172,67 @@ where
   R: Fn(&O::Value) -> &V + Clone,
   W: Fn(&mut O::Value) -> &mut V + Clone,
 {
-  splitted_reader_impl!();
+  /// set your custom checker to replace the default checker,
+  /// See[`VolatileSplit`].
+  /// ### Example of set_checker to SplittedWriter
+  /// ``` rust
+  /// use ribir_core::prelude::*;
+  /// #[derive(Default)]
+  /// struct App {
+  ///    name: String,
+  ///    data: Vec<String>,
+  ///    //...
+  /// }
+  /// let app = Stateful::new(App::default());
+  /// let data = app.split_writer(|a| &a.data, |a| &mut a.data).set_checker(AlwaysValidSplit {});
+  /// app.write().name = "ribir".to_string();
+  ///
+  /// // the use of data will cause panic if not set the checker to AlwaysValidSplit. See [`VolatileSplit`]
+  /// println!("data len {}", data.read().len());
+  /// ```
+  pub fn set_checker<C2>(self, checker: C2) -> SplittedWriter<O::Writer, R, W, C2>
+  where
+    C2: SplitChecker<O::Value>,
+  {
+    assert!(
+      self.ref_count.get() == 1,
+      "SplittedWriter's set_checker should be call after create and before used"
+    );
+    SplittedWriter {
+      origin: self.origin.clone_writer(),
+      map: self.map.clone(),
+      mut_map: self.mut_map.clone(),
+      notifier: self.notifier.clone(),
+      batched_modify: self.batched_modify.clone(),
+      checker: Sc::new(checker),
+      last_modified: self.last_modified.clone(),
+      ref_count: self.ref_count.clone(),
+    }
+  }
 }
 
-impl<V, O, R, W> StateWriter for SplittedWriter<O, R, W>
+impl<C, V, O, R, W> StateReader for SplittedWriter<O, R, W, C>
 where
   Self: 'static,
   V: ?Sized,
   O: StateWriter,
   R: Fn(&O::Value) -> &V + Clone,
   W: Fn(&mut O::Value) -> &mut V + Clone,
+  C: SplitChecker<O::Value>,
 {
-  type Writer = SplittedWriter<O::Writer, R, W>;
+  splitted_reader_impl!();
+}
+
+impl<V, O, R, W, C> StateWriter for SplittedWriter<O, R, W, C>
+where
+  Self: 'static,
+  V: ?Sized,
+  O: StateWriter,
+  R: Fn(&O::Value) -> &V + Clone,
+  W: Fn(&mut O::Value) -> &mut V + Clone,
+  C: SplitChecker<O::Value>,
+{
+  type Writer = SplittedWriter<O::Writer, R, W, C>;
   type OriginWriter = O;
 
   #[inline]
@@ -151,8 +251,8 @@ where
       mut_map: self.mut_map.clone(),
       notifier: self.notifier.clone(),
       batched_modify: self.batched_modify.clone(),
+      checker: self.checker.clone(),
       last_modified: self.last_modified.clone(),
-      create_at: self.create_at,
       ref_count: self.ref_count.clone(),
     }
   }
@@ -161,12 +261,13 @@ where
   fn origin_writer(&self) -> &Self::OriginWriter { &self.origin }
 }
 
-impl<V, O, R, W> WriterControl for SplittedWriter<O, R, W>
+impl<V, O, R, W, C> WriterControl for SplittedWriter<O, R, W, C>
 where
   Self: 'static,
   V: ?Sized,
   O: StateWriter,
   R: Fn(&O::Value) -> &V + Clone,
+  C: SplitChecker<O::Value>,
   W: Fn(&mut O::Value) -> &mut V + Clone,
 {
   #[inline]
@@ -182,36 +283,32 @@ where
   fn dyn_clone(&self) -> Box<dyn WriterControl> { Box::new(self.clone_writer()) }
 }
 
-impl<V, O, R, W> SplittedWriter<O, R, W>
+impl<V, O, R, W, C> SplittedWriter<O, R, W, C>
 where
   Self: 'static,
   O: StateWriter,
   R: Fn(&O::Value) -> &V + Clone,
   W: Fn(&mut O::Value) -> &mut V + Clone,
+  C: SplitChecker<O::Value>,
   V: ?Sized,
 {
-  pub(super) fn new(origin: O, map: R, mut_map: W) -> Self {
+  pub(super) fn new(origin: O, map: R, mut_map: W, checker: C) -> Self {
     let create_at = Instant::now();
-
     Self {
       origin,
       map,
       mut_map,
       notifier: Notifier::default(),
       batched_modify: <_>::default(),
+      checker: Sc::new(checker),
       last_modified: Sc::new(Cell::new(create_at)),
-      create_at,
       ref_count: Sc::new(Cell::new(1)),
     }
   }
 
   #[track_caller]
   fn split_ref<'a>(&'a self, mut orig: WriteRef<'a, O::Value>) -> WriteRef<'a, V> {
-    assert!(
-      self.create_at > self.origin.time_stamp(),
-      "A splitted writer is invalid because its origin state is modified after it created."
-    );
-
+    assert!(self.is_valid(), "A splitted writer is invalid.");
     let modify_scope = orig.modify_scope;
 
     // the origin mark as a silent write, because split writer not effect the origin
@@ -234,11 +331,12 @@ where
   }
 }
 
-impl<V, O, R, W> RenderBuilder for SplittedWriter<O, R, W>
+impl<V, O, R, W, C> RenderBuilder for SplittedWriter<O, R, W, C>
 where
   O: StateWriter,
   R: Fn(&O::Value) -> &V + Clone + 'static,
   W: Fn(&mut O::Value) -> &mut V + Clone,
+  C: SplitChecker<O::Value>,
   V: Render,
 {
   fn widget_build(self, ctx: &BuildCtx) -> Widget {